Spels
-Добавлен скрипт SpelDiscription, содержащий в себе информацию о всех настройках спелов (Форма и дальность спела, можете потыкать, посмотреть (CastRange пока не работает), на кого направлен скрипт(onGround пока ни на что не влияет), его цели, урон и хил(Для зданий не работает большая часть процентных изменений), условия (кол-во действий, чтобы скастовать, кол-во действий, которые потратятся, IsPassive и IsCalling пока ни на что не влияют, IsOnButton нужно, чтобы появилась кнопка его каста), эффектов у заклинаний пока нет)
-Добавлен скрипт Caster, содержащий в себе всю логику каста заклинаний (Взаимодействие заклинаний с водой пока не прописано, перепады высот игнорируются)
-Сделан кастомный инспектор для скрипта Caster
-Добавлена UI панель, появляющаяся при выделении юнита или здания, отображающая имя, здоровье и кнопочки спелов, если они есть(чтобы скастовать спел, нужно нажать на кнопку этого спела, а затем на клетку на поле, сейчас все клетки, которые покрывает спел, для наглядности подкрашиваются в красный( ps клетка, на которую вы кликнули тоже в зоне действия спела, но не подкрашивается из-за её выделения))
-Добавлен префаб SelectedObjectInformationEnableController(его нужно вытащить на сцену), который отвечает за то, чтобы появлялась панель из пункта выше
-В TurnManager добавлена ф-ция SetChangesInOtherClassesOnTurnChanged(), в которой можно будет указывать, что должны сделать другие скрипты после смены хода(преимущество перед экшеном тут в том, что можно самим задать порядок выполнения действий)
-В UnitHealth добавлены новые ф-ции для разных вариантов убавления здоровья(Такие же ф-ции надо добавить зданиям)
-Добавлен префаб первого юнита второй фракции Warlock
-В MouseSelection раскомментированна строчка, отвечающая за то, чтобы выделение клеток не работало, если мышка на UI элементе
-Рефакторинг кода

-Ещё есть, над чем работать, не уверен, что всё проверил, закоммитил сейчас, чтобы вы могли ознакомиться с механикой спелов, возможно, предложить что-нибудь. Если будут вопросы или найдёте баги, пишите
-ps на полностью закомментированные скрипты в папке Editor не обращайте внимания)

+++ Assets/Scripts/HexGrid.cs
@@ -31,7 +31,7 @@ public class HexGrid : MonoBehaviour {
     * f e *
     * * * *
     */
-    [SerializeField] private int spacing;
+    [SerializeField] private float spacing;
 
     [Header("Terrain Generation")]
     [SerializeField] private Noise baseHeight;
